Fix rendering of self-closing cubic bezier curve (fix #44).
Previously, if the curve was self closing, it was up to numerical
error for it to get rendered, it could be declared as sub-pixel
nearly immediatly. Now a global check look at this particular case,
break the curve in two non-self closing curve and renders them.

Not incredible but does the work.